# docs/swagger.yaml
definitions:
...
  feedback.FeedbackInput:
    properties:
      email:
        type: string
      text:
        type: string
    required:
    - email
    - text
    type: object
...
info:
  contact: {}
paths:
...
  /event:
    get:
      consumes:
      - application/json
      description: Dapatkan semua event atau event berdasarkan ID tertentu
      parameters:
      - description: ID Event
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan semua event atau event berdasarkan ID tertentu
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: Buat event baru dengan informasi yang diberikan
      parameters:
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Judul
        in: formData
        name: judul
        required: true
        type: string
      - description: Pesan Event
        in: formData
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat event baru
      tags:
      - Event
  /event/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus event berdasarkan ID yang diberikan
      parameters:
      - description: ID Event
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus event berdasarkan ID
      tags:
      - Event
    get:
      consumes:
      - application/json
      description: Dapatkan satu event berdasarkan ID yang diberikan
      parameters:
      - description: ID Event
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan satu event berdasarkan ID
      tags:
      - Event
    put:
      consumes:
      - application/json
      description: Update event berdasarkan ID yang diberikan
      parameters:
      - description: ID Event
        in: path
        name: id
        required: true
        type: integer
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Judul
        in: formData
        name: judul
        required: true
        type: string
      - description: Pesan Event
        in: formData
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update event berdasarkan ID
      tags:
      - Event
  /feedback:
    get:
      consumes:
      - application/json
      description: Dapatkan semua feedback atau feedback berdasarkan ID tertentu
      parameters:
      - description: ID Feedback
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan semua feedback atau feedback berdasarkan ID tertentu
      tags:
      - Feedback
    post:
      consumes:
      - application/json
      description: Buat feedback baru dengan informasi yang diberikan
      parameters:
      - description: Data Feedback yang ingin dibuat
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/feedback.FeedbackInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Buat feedback baru
      tags:
      - Feedback
  /feedback/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus feedback berdasarkan ID yang diberikan
      parameters:
      - description: ID Feedback
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus feedback berdasarkan ID
      tags:
      - Feedback
    get:
      consumes:
      - application/json
      description: Dapatkan feedback berdasarkan ID yang diberikan
      parameters:
      - description: ID Feedback
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan feedback berdasarkan ID
      tags:
      - Feedback
...
swagger: "2.0"